Local development: Inital UI architecture
