+++ package.json
@@ -8,19 +8,21 @@
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "npm": "^10.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-pdf": "^5.3.0",
-    "react-scripts": "4.0.3",
+    "react-pdf": "^9.0.0",
+    "react-scripts": "^4.0.3",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
+    "deploy": "gh-pages -d dist",
+    "build": "tsc",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "prettier --write 'src/**/*.{ts,tsx}'"
   },
   "eslintConfig": {
     "extends": [
@@ -41,6 +43,10 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.0"
+    "eslint-config-prettier": "^9.1.0",
+    "gh-pages": "^6.1.1",
+    "http-server": "^0.12.3",
+    "prettier": "^3.3.2",
+    "typescript": "^4.9.5"
   }
 }
